Fix Qt Wayland Compositor crash on Wayland 1.23
With Wayland 1.23, Qt Wayland Compositor is running into trouble
because the waylandscanner generates public-code. In Qt 6.8 this
has been fixed by enabling the cmake API to support both
private-code and public-code with an argument, and running it with
private-code for Qt sources.

However, since this is public API, we need a hotfix in Qt 6.7 and
before. This patch introduces an undocumented cmake function which
passes private-code to waylandscanner and uses this for Qt sources.

Pick-to: 6.2
Task-number: QTBUG-126301
Change-Id: I7d6a61199ff22b68d9b7dce09262a0079e3be3da
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
(cherry picked from commit cdae5c64fb11466f5d489add1ffb645ae5674fed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>