Avoid calling handleScreenRemoved() recursively
In some cases where a nested event loop is run as a result
of a screen being removed (for instance if it's called from
primaryScreenChanged() or similar signals) and multiple
screens were added to the system, then we could end
up calling handleScreenRemoved() for the placeholder screen
recursively. This  would cause a crash if anyone actually
used the QScreen pointer passed through the screenRemoved()
signal.

This makes sure we set the placeholder screen to null before
calling handleScreenRemoved() so that we don't call it again
if we happen to end up in the same function in a nested
event loop.

Pick-to: 6.8 6.5
Change-Id: I237946851ed4dce03fd53093baba102c9686be46
Reviewed-by: David Edmundson <davidedmundson@kde.org>
(cherry picked from commit e038d0b750cf238fef6eeb07a90b7fc1b52e4640)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>