Fix compiling cpp-client example
Including the QQmlEngine header in the shared client code
produced debugging code that depended on QtQml. Since this
is not linked from the cpp-client, it would fail to compile.

The only thing we need is the QML_ELEMENT macro, and for
that we can just include QtQmlIntegration instead, which
does not impose the dependency.

Pick-to: 6.8
Fixes: QTBUG-132642
Change-Id: Iec3bd99404edb166c7db847ead95198627523379
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 24735372c5fbf0d3c0ca9a3624ad156632dfb244)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>