client: Drop assert in frame size syncing
Whilst this assert is true for a normal application workflow
such as QtQuick, it's not something we can guarantee with all
applications.

If the assert is not true, the absolutely worst case is the
same glitchy resizing that we had in 6.8 and older.

Fixes: QTBUG-134126
Pick-to: 6.9.0
Pick-to: 6.9.0
Change-Id: I43b660d771d628ceedce9407fd146b48074557f9
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 9e971b9fcf2ebce2639384b72498ac0b01ba1c84)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>