compositor: Avoid overflow for infiniteRegion()
The infinite region is just supposed to be an arbitrarily large rectangle that covers everything, but since it's put into a QRegion, its width will be calculated.

Therefore the width has to fit inside an int. We adjust the bounds so that its width becomes exactly numerical_limits<int>::max().

This is caught by 419ebb2770eb606762dcca96978b7b26299aafd9 and the
series in qtbase.

Fixes: QTBUG-134939
Change-Id: I03c9ebbba0288594f87a131b0dfc28a104c8c91d
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>