Client: Reset mFrameCallbackTimedOut when showing a window
When a window is hidden, the corresponding expose event is sent
asynchronously. So, a buffer (and a frame callback) can be attached
after the window is hidden.

QWaylandWindow::initWindow() handles that case by attaching a nil buffer
to the surface, but it skips resetting the frame callback things.

If mFrameCallbackTimedOut is not reset, then the window can get stuck
in the unexposed state and not show up after the second setVisible(true).

This change makes the initWindow() function also reset frame callback
data. It fixes some windows not showing up after remapping them.

Pick-to: 6.9
Change-Id: I8952ffed24844c73721c4f689786146c317e014d
Reviewed-by: David Edmundson <davidedmundson@kde.org>