Client: Fix copying shm buffer data with fractional scaling
The QWaylandShmBackingStore computes the device pixels in the source
image manually. But it does not match what the QPainter does one to one.
Visually, it produces artifacts where the contents of an app is
jiggling or moves back and forth by one pixel.

In order to address the issue, the QWaylandShmBackingStore can copy
whole buffers but with a clip region set. That way, the QPainter can
consistently compute the coordinates in the device pixels.

Pick-to: 6.9
Change-Id: Ib4832c4d687f43f6cc1de4a1760c32f1abd2ab03
Reviewed-by: David Edmundson <davidedmundson@kde.org>