qttextinput: Fix deadlock between client and compositor
When sending the end_input_method event the compositor waits
until the event is confirmed. This is done by polling wayland
events in a loop.

If the client sends a new frame at the same time the compositor
sends a the end_input_method event, the client cannot answer
wayland events as it's waiting in swapBuffers until the frame
callback was called. The compositor on the other hand just polls
new wayland events.

By making sure to also process normal Qt wayland events, we make
sure the scene graph is not blocked and we can send the frame
callbacks to unblock the client.

Pick-to: 6.5 6.8 6.9 6.10
Change-Id: I328710193008728e4eac2bbc4e6eba9c81567d4d
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>