QWaylandAdwaitaDecoration: unify window border into a single path
We originally painted the decoration by splitting it to two parts. First
part was the titlebar with rounded corners and second part was the rest
of the window, basically to paint borders around the window. This had
a side-effect, where part of the border was overpainted and causes an
artifact when a transparent border is used.

Pick-to: 6.9
Change-Id: If7011c9c99e04e2362c8491c08cc7ba82b849378
Reviewed-by: David Edmundson <davidedmundson@kde.org>
(cherry picked from commit d0a893df1f59a79799fa69403230d53035aec166)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>