WaylandTextureSharing: remove use of NO_GENERATE_QMLTYPES
Remove the use of NO_GENERATE_QMLTYPES in WaylandTextureSharing, where
it seems that the plugin does not register any types, and in
WaylandTextureSharingExtension, where one type gets type registered
declaratively via QML_FOREIGN.

Add an extra foreign types for the base types of
QWaylandTextureSharingExtensionQuickExtension to avoid qmllint warnings
about unresolved types. It seems to be a drawback from the
Q_COMPOSITOR_DECLARE_QUICK_EXTENSION_NAMED_ELEMENT macro, that only
registers a type but not all of the base classes of a type.

Task-number: QTBUG-110794
Change-Id: Ic5e2bcf95f6029422dd94deda62d158b53387a70
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>