Bump wl_compositor support to v5
For wl_surface version is 5 or higher, passing any non-zero
x or y to wl_surface.attach is a protocol violation,
and will result in an 'invalid_offset' error being raised.
The x and y arguments are ignored and do not change the pending state.
To achieve equivalent semantics, use wl_surface.offset.

Change-Id: I70f72c747d8f8666e1437050d4b0e9227c439f6d
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>