Consolidate done events on outputs
The XDG specific done event was deprecated in favor of the one
on wl_output, but the logic to select which one to send was
duplicated in multiple places and in the case of bind, we were
not applying the right logic. This causes some clients to ignore
important properties, with a subsequent crash.

This patch centralizes the logic and adds the call to
wl_output.done to the bind code path as well.

Pick-to: 6.8 6.10 6.11
Fixes: QTBUG-142162
Change-Id: If501a641dc8fc2d828f24492953244ad4e60e8ae
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Ilya Fedin <fedin-ilja2010@ya.ru>